Addressing PR feedback: extract command processing to separate file, create `ui/noniteractive` directory
